+++ src/mgfw/math/Rect.h
@@ -41,8 +41,8 @@ struct Rect {
 
     /* Operators */
 
-    T& operator[](uint16_t index) { return data[ index ]; }
-    const T& operator[](uint16_t index) const { return data[ index ]; }
+    T& operator[](uint8_t index) { return data[ index ]; }
+    const T& operator[](uint8_t index) const { return data[ index ]; }
 
     template <typename U>
     bool operator==(const Rect<U>& other) const {
