+++ src/mgfw/math/Vector2.h
@@ -26,8 +26,8 @@ struct Vector2 {
 
     /** Operators */
 
-    T& operator[](uint16_t index) { return data[index]; }
-    const T& operator[](uint16_t index) const { return data[index]; }
+    T& operator[](uint8_t index) { return data[index]; }
+    const T& operator[](uint8_t index) const { return data[index]; }
 
     template <typename U>
     bool operator==(const Vector2<U>& other) const {
